Add Clang flag -fno-discard-value-names for better LLVM IR output.
The Godbolt versions of Clang are built w/o assertions enabled,
this means that Clang discards value and label names when generating
LLVM IR. However, since the point of Godbolt is to inspect the assembly
or, often, the LLVM IR, it's useful to have value names present.

This patch adds the newely added -fno-discard-value-names option to the
default Clang trunk options. This prevents value names from being
discarded and creates more readable LLVM IR.

The flag was added to Clang today (Feb 7th, 2018), so it's
only applicable to trunk builds.